@@ -16,17 +16,19 @@
     "skeleton"
   ],
   "dependencies": {
-    "@material-ui/core": "^3.6.2",
-    "@material-ui/icons": "^3.0.1",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/material": "^5.9.0",
+    "@reduxjs/toolkit": "^1.8.1",
     "axios": "^0.18.0",
     "chalk": "^2.4.1",
     "express": "^4.14.0",
     "nodemon": "^2.0.4",
     "pg": "^8.7.3",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-redux": "^5.0.7",
-    "react-router-dom": "^4.2.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-redux": "^v8.0.1",
+    "react-router-dom": "^6.3.0",
     "redux": "^4.0.0",
     "redux-devtools-extension": "^2.13.7",
     "redux-logger": "^3.0.6",
@@ -34,20 +36,19 @@
     "sequelize": "^6.16.1"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.3",
-    "babel-loader": "^7.1.4",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
+    "@babel/core": "7.17.10",
+    "@babel/polyfill": "^7.12.1",
+    "@babel/preset-env": "7.17.10",
+    "@babel/preset-react": "7.16.7",
+    "@babel/preset-stage-2": "^7.8.3",
+    "babel-loader": "8.2.5",
     "css-loader": "^3.0.0",
     "eslint": "^4.19.1",
     "eslint-config-fullstack": "^5.0.0",
     "eslint-plugin-react": "^7.8.2",
     "style-loader": "^0.21.0",
     "volleyball": "^1.5.1",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.3.0"
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2"
   }
 }
