@@ -1,33 +1,22 @@
 import React from 'react';
-import {Link} from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
-import CardActionArea from '@material-ui/core/CardActionArea';
-import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
-import Typography from '@material-ui/core/Typography';
+import {
+  Card,
+  CardActionArea,
+  CardContent,
+  CardMedia,
+  Typography,
+} from '@mui/material';
 
-const styles = {
-  card: {
-    maxWidth: 500,
-  },
-  media: {
-    objectFit: 'cover',
-  },
-};
-
-const Candy = (props) => {
-  const {candy, classes} = props;
-
+const Candy = ({ candy }) => {
   return (
     <Link to={`/candies/${candy.id}`}>
-      <Card className={classes.card}>
+      <Card className="candy-card">
         <CardActionArea>
           <CardMedia
             component="img"
             alt={`${candy.name} candy image`}
-            className={classes.media}
             height="140"
             image={candy.imageUrl}
             title={`${candy.name} candy image`}
@@ -40,11 +29,7 @@ const Candy = (props) => {
         </CardActionArea>
       </Card>
     </Link>
-  )
-}
-
-Candy.propTypes = {
-  classes: PropTypes.object.isRequired
+  );
 };
 
-export default withStyles(styles)(Candy);
+export default Candy;
