@@ -1,33 +1,30 @@
-'use strict'
+'use strict';
 
-const { resolve } = require('path')
+const { resolve } = require('path');
 
 module.exports = {
-  entry: ['babel-polyfill', './app/main'],
+  entry: ['@babel/polyfill', './app/main'],
   output: {
     path: __dirname,
-    filename: './public/bundle.js'
+    filename: './public/bundle.js',
   },
   mode: 'development',
   context: __dirname,
   devtool: 'source-map',
   resolve: {
-    extensions: ['.js', '.jsx']
+    extensions: ['.js', '.jsx'],
   },
   module: {
     rules: [
       {
         test: /jsx?$/,
         include: resolve(__dirname, './app'),
-        loader: 'babel-loader'
+        loader: 'babel-loader',
       },
       {
         test: /\.css$/,
-        use: [
-          'style-loader',
-          'css-loader'
-        ]
-      }
-    ]
-  }
-}
+        use: ['style-loader', 'css-loader'],
+      },
+    ],
+  },
+};
